+++ src/icemac/recurrence/__init__.py
@@ -1,1 +1,1 @@
-#
+from .recurrence import get_recurrences, get_recurring  # noqa