+++ CHANGES.rst
@@ -5,7 +5,13 @@
 1.0.2 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Shorten the import path of ``get_recurrences()`` from
+  ``icemac.recurrence.recurrence`` to just ``icemac.recurrence``.
+
+- Refactor tests to use ``py.test`` fixtures.
+
+- Fix an off by one month error in the periods `nth weekday every other month`
+  and `nth weekday from end of other month`.
 
 
 1.0.1 (2015-10-22)